+++ CHANGELOG.md
@@ -7,11 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2021-09-19
+### Added
+- config was added to support configuration errors
+### Removed
+- platform error type was removed, not needed
+
 ## [0.2.0] - 2021-08-21
 ### Added
 - defer error for use inside defer functions
 
 ## [0.1.0] - 2021-08-21
-### Added
 - platform, system and server failures
 - not found, validation, input and ignore failures